backend/requirements.txt
asgiref==3.7.2
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==4.2.8
django-allauth==0.61.1
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
environ==1.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
pillow==11.0.0
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
